FASTAPI/requirements.txt
fastapi
uvicorn
requests
torch
transformers
pydantic
python-dotenv
google-generativeai
nltk